refactor(runtime): split lifecycle supervision
Findings: AO2-ARCH-001, AO2-QA-004, AO2-QA-005
