+++ crates/atm-daemon-bootstrap/src/lib.rs
@@ -221,10 +221,47 @@ async fn run_replacement_daemon_with_selector(
     let runtime_health = RuntimeHealth::with_owner(std::process::id());
     let assembly = assemble_daemon_runtime()?;
     let workflow_telemetry = assembly.workflow_telemetry.clone();
-    // AO.2 composes concrete TLS only at the daemon bootstrap boundary. The
-    // runtime receives this sealed, opaque core adapter and never observes a
-    // certificate, peer store, or Rustls type.
-    let peer_io_adapter = match peer_tls::mtls_adapter(assembly.peer_config_store()) {
+    let peer_io_adapter = optional_peer_io_adapter(&assembly);
+    let mut running = start_replacement_runtime(
+        &scope,
+        &_owner,
+        assembly,
+        observability,
+        peer_io_adapter,
+        selector_factory,
+        &daemon_launch_identity,
+        runtime_health,
+    )
+    .await?;
+    if let Err(error) = emit_ready_signal_if_requested() {
+        // The process has not advertised readiness, so it must not retain an
+        // otherwise-live listener when its supervisor handshake fails.
+        let _ = running.begin_shutdown().finish().await;
+        workflow_telemetry.shutdown().await;
+        return Err(error);
+    }
+    if let ShutdownTrigger::ServerStopped = wait_for_replacement_shutdown(&mut running).await? {
+        let result = match running.begin_shutdown().finish().await {
+            Ok(_) => Err(AtmError::daemon_unavailable(
+                "replacement HTTP runtime server stopped unexpectedly",
+            )),
+            Err(error) => Err(error),
+        };
+        workflow_telemetry.shutdown().await;
+        return result;
+    }
+    let _stopped = running.begin_shutdown().finish().await?;
+    workflow_telemetry.shutdown().await;
+    Ok(())
+}
+
+/// AO.2 composes concrete TLS only at the daemon bootstrap boundary. The
+/// runtime receives this sealed, opaque core adapter and never observes a
+/// certificate, peer store, or Rustls type.
+fn optional_peer_io_adapter(
+    assembly: &RuntimeAssembly,
+) -> Option<Arc<dyn atm_core::PeerIoAdapter>> {
+    match peer_tls::mtls_adapter(assembly.peer_config_store()) {
         Ok(adapter) => Some(adapter),
         Err(error) => {
             // mTLS is optional until a valid exchange configuration exists.
@@ -233,7 +270,21 @@ async fn run_replacement_daemon_with_selector(
             tracing::info!(error = %error, "direct peer mTLS is unavailable; listener remains disabled");
             None
         }
-    };
+    }
+}
+
+async fn start_replacement_runtime(
+    scope: &atm_core::home::HostRuntimeScope,
+    owner: &DaemonOwnerGuard,
+    assembly: RuntimeAssembly,
+    observability: Arc<dyn ObservabilityPort + Send + Sync>,
+    peer_io_adapter: Option<Arc<dyn atm_core::PeerIoAdapter>>,
+    selector_factory: impl FnOnce(
+        atm_core::LocalServiceRuntime,
+    ) -> Arc<dyn atm_core::boundary::MessageReceivedHookSelector>,
+    daemon_launch_identity: &DaemonLaunchIdentity,
+    runtime_health: RuntimeHealth,
+) -> Result<atm_http_runtime::HttpRuntime<atm_http_runtime::Running>, AtmError> {
     // The shipped daemon always keeps the injected receiver hook active.
     // Benchmark-only selection is available only from the separate binary.
     let handler = build_replacement_handler(
@@ -244,10 +295,24 @@ async fn run_replacement_daemon_with_selector(
         &daemon_launch_identity,
         runtime_health.clone(),
     )?;
+    let config = replacement_runtime_config(scope, owner, peer_io_adapter.is_some())?;
+    let builder = HttpRuntimeBuilder::new(config, handler).with_runtime_health(runtime_health);
+    let builder = match peer_io_adapter {
+        Some(peer_io_adapter) => builder.with_peer_io_adapter(peer_io_adapter),
+        None => builder,
+    };
+    builder.build()?.start().await
+}
+
+fn replacement_runtime_config(
+    scope: &atm_core::home::HostRuntimeScope,
+    owner: &DaemonOwnerGuard,
+    peer_transport_enabled: bool,
+) -> Result<HttpRuntimeConfig, AtmError> {
     let loopback = LoopbackTcpConfig::new(
         SocketAddr::new(IpAddr::V4(Ipv4Addr::LOCALHOST), 0),
         scope.runtime_root.as_ref().join(LOCAL_HTTP_RECORD_FILENAME),
-        _owner.instance_id(),
+        owner.instance_id(),
     );
     let config = HttpRuntimeConfig::new(
         loopback,
@@ -262,45 +327,32 @@ async fn run_replacement_daemon_with_selector(
             NonZeroDuration::new(REPLACEMENT_DRAIN_DEADLINE).expect("non-zero shutdown timeout"),
         ),
     );
-    let config = if peer_io_adapter.is_some() {
+    Ok(if peer_transport_enabled {
         config.with_direct_peer_tcp(DirectPeerTcpConfig::standard())
     } else {
         config
-    };
-    let builder = HttpRuntimeBuilder::new(config, handler).with_runtime_health(runtime_health);
-    let builder = match peer_io_adapter {
-        Some(peer_io_adapter) => builder.with_peer_io_adapter(peer_io_adapter),
-        None => builder,
-    };
-    let mut running = builder.build()?.start().await?;
-    if let Err(error) = emit_ready_signal_if_requested() {
-        // The process has not advertised readiness, so it must not retain an
-        // otherwise-live listener when its supervisor handshake fails.
-        let _ = running.begin_shutdown().finish().await;
-        workflow_telemetry.shutdown().await;
-        return Err(error);
-    }
+    })
+}
+
+enum ShutdownTrigger {
+    Signal,
+    ServerStopped,
+}
+
+async fn wait_for_replacement_shutdown(
+    running: &mut atm_http_runtime::HttpRuntime<atm_http_runtime::Running>,
+) -> Result<ShutdownTrigger, AtmError> {
     tokio::select! {
         signal = wait_for_shutdown_signal() => {
             let signal = signal?;
             eprintln!("replacement ATM daemon received {}; starting graceful shutdown", signal.as_str());
+            Ok(ShutdownTrigger::Signal)
         }
         _ = running.wait_for_server_stop() => {
             eprintln!("replacement ATM HTTP runtime server stopped unexpectedly; beginning cleanup");
-            let result = match running.begin_shutdown().finish().await {
-                Ok(_) => Err(AtmError::daemon_unavailable(
-                    "replacement HTTP runtime server stopped unexpectedly",
-                )),
-                Err(error) => Err(error),
-            };
-            workflow_telemetry.shutdown().await;
-            return result;
+            Ok(ShutdownTrigger::ServerStopped)
         }
     }
-    let stopped = running.begin_shutdown().finish().await;
-    workflow_telemetry.shutdown().await;
-    let _stopped = stopped?;
-    Ok(())
 }
 
 /// Emit the benchmark/supervisor marker only after every enabled replacement
